requirements.txt
groq
pdfplumber
pandas
pytest
python-dotenv